add authentication

+++ graphql/resolvers/booking.js
@@ -4,7 +4,10 @@ const { transformEvent, transformBooking } = require('./merge')
 
 module.exports = {
 
-    bookings: async () => {
+    bookings: async (args, req) => {
+        if (!req.isAuth) {
+            throw new Error('Unauthenticated!');
+        }
         try {
             const bookings = await Booking.find();
             return bookings.map(booking => {
@@ -14,7 +17,10 @@ module.exports = {
             throw err;
         }
     },
-    bookEvent: async args => {
+    bookEvent: async (args, req) => {
+        if (!req.isAuth) {
+            throw new Error('Unauthenticated!');
+        }
         const fetchedEvent = await Event.findOne({ _id: args.eventId });
         const booking = new Booking({
             user: '6129a29d23ecd562449ddb5b',
@@ -23,7 +29,10 @@ module.exports = {
         const result = await booking.save();
         return transformBooking(result);
     },
-    cancelBooking: async args => {
+    cancelBooking: async (args, req) => {
+        if (!req.isAuth) {
+            throw new Error('Unauthenticated!');
+        }
         try {
             const booking = await Booking.deleteOne({ _id: args.bookingId });
             const event = transformEvent(booking.event)
